info: forward all the information we receive
I'm a fool and I had only encoded the information I needed for my own
project (the email in that case), which means everything else was
thrown out. Now the `info` hash contains everything from the PC claims[1]
that can be mapped to the OmniAuth Auth Schema[2] (because info must
remain standardized), and put everything else in the extra hash.

[1]: https://partenaires.proconnect.gouv.fr/docs/fournisseur-service/scope-claims
[2]: https://github.com/omniauth/omniauth/wiki/Auth-Hash-Schema